Fixed a path concatenation bug in `retrieval_configuration`
I found a case in `retrieval_configuration` where it concatenated a default directory
and a file path with string interpolation instead of path operations, which meant that
if the file path was already an absolute path, you end up with an incorrect prepending
of the default path to the absolute path. Switching to os.path.join fixes this.