Update handling of SWIR ILS and monochromatic grid
There were two coupled problems in the SWIR ILS. First, I was using the old
py-retrieve code to load the ILS parameters, which pre-interpolates the ILS
to all the monochromatic frequencies. This causes the ReFRACtor ILS table
to have to represent all of those, which isn't necessary. Second, that also
meant that the ReFRACtor interface using the ILS central wavelengths as the
monochromatic wavelengths would no longer work, so I've modified that to
use the ABSCO grid instead.